Add an optional columns property to checkboxSet
We want to be able to render a checkbox set horizontally. In order to do
so, we need to be able to specify the number of columns to be used
during rendering as we do for tabular select.

Since `Formlet.Ocelot.Checkbox.Render` implements `Semigroup`, `columns`
must be a `Monoid`. While we could select any arbitrary monoid (such as
`First Int`), we decide to use `Array Int` to preserve values and
ordering. While it's unlikely any `render` function will make use of
this sequence, we shouldn't make assumptions.