remove `id` from Toggle input
We have `for` that links the same `id` on the form field label/title so when user
clicks on the title, the toggle switches. Per Product request, we need to remove
this behavior to be consistent with the rest of the site so we remove the `id`
here to de-link the title and the toggle input.